Follow-up
- Fix clipboard definitions
- Add Menu & Palette item types to protocol
