+++ packages/client/src/index.ts
@@ -80,7 +80,6 @@ export * from './features/tool-feedback/css-feedback';
 export * from './features/tool-feedback/edge-edit-tool-feedback';
 export * from './features/tool-feedback/feedback-action-dispatcher';
 export * from './features/tool-feedback/model';
-export * from './features/tool-palette/palette-item';
 export * from './features/tool-palette/tool-palette';
 export * from './features/tools/base-glsp-tool';
 export * from './features/tools/change-bounds-tool';
